Remove Stats Bloc and move stats files to model
