+++ lib/blocs/blocs.dart
@@ -6,7 +6,6 @@ export 'src/filtered_refuelings/barrel.dart';
 export 'src/filtered_todos/barrel.dart';
 export 'src/notifications/barrel.dart';
 export 'src/refuelings/barrel.dart';
-export 'src/stats/barrel.dart';
 export 'src/tab/barrel.dart';
 export 'src/todos/barrel.dart';
 export 'src/database/barrel.dart';
