+++ lib/models/models.dart
@@ -12,3 +12,5 @@ export 'src/repeat.dart';
 export 'src/distanceratepoint.dart';
 export 'src/distancepoint.dart';
 export 'src/fuel_mileage_point.dart';
+export 'src/stats/driving_distance.dart';
+export 'src/stats/efficiency.dart';
